+++ index.js
@@ -1,7 +1,0 @@
-import def from './cjs/index.js';
-
-export const random = def.random;
-export const randomFloat = def.randomFloat;
-export const randomFromArray = def.randomFromArray;
-export const randomFloats = def.randomFloats;
-export const randomIntegers = def.randomIntegers;