+++ index.mjs
@@ -1,0 +1,2 @@
+import cjsModule from './cjs/index.js';
+export const { __esModule, random, randomFloat, randomFromArray, DEFAULT_RANDOM_NUMBERS_MAX, randomIntegers, randomFloats, LOWERS, UPPERS, LETTERS, randomLower, randomUpper, randomLetter, randomLowers, randomUppers, randomLetters, randomDate, randomDates } = cjsModule;